STEP 2 : OCP의 실현, 성능 리팩터링
