+++ WeatherForecast/WeatherForecast/Model/Weather.swift
@@ -1,0 +1,36 @@
+//
+//  WeatherForecast - Weather.swift
+//  Created by yagom.
+//  Copyright © yagom. All rights reserved.
+// 
+
+import Foundation
+
+// MARK: - Weather JSON Format
+struct WeatherJSON: Decodable {
+    static let changeWeaterJsonKey: String = "changeWeatherJson"
+    let weatherForecast: [WeatherForecastInfo]
+    let city: City
+}
+
+// MARK: - List
+struct WeatherForecastInfo: Decodable {
+    let dt: TimeInterval
+    let main: MainInfo
+    let weather: Weather
+    let dtTxt: String
+}
+
+// MARK: - MainClass
+struct MainInfo: Decodable {
+    let temp, feelsLike, tempMin, tempMax: Double
+    let pressure, seaLevel, grndLevel, humidity, pop: Double
+}
+
+// MARK: - Weather
+struct Weather: Decodable {
+    let id: Int
+    let main: String
+    let description: String
+    let icon: String
+}
